zephyr: extract heap functionality into a separate file
Create alloc.c for all heap management code.
